Vector database Assignment_4/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4